Properly memoize the initialTarget in useRefs
The memoized initialTarget in useRefs was thrown away when the
reference was changed. This is not the expected behavior as the
initial value should be created just once.

It's not possible to create a test for this because the hook does not
expose the initial reference through the proxy. The change was manually
tested.