Make arrayRef stable between re-renders
Previously, arrayRef always returned a new ref-function when being
executed in a render. Components that receive this ref function might
execute logic when this ref changes (e.g. the ensureForwardRef HoC).
This could result in too much re-renders, or in some cases infinite
recursive rendering where parents and children keep updating each other.

This change applies a memoize to the arrayRef implementation. It does
this by generating a unique key for each ref/index combination
(the function's parameters). Internally it uses a Map to allow mapping
the ref objects to random strings, that can in turn be used as keys.